fix upgrade to pass from 18.03 to > 19.03
Delete `22_entity_tenant_migration.py` because migration will be done
by alembic scripts.

We stopped the alembic scripts after adding `tenant_uuid` column in
`entity` table, to insert tenant in wazo-auth and fix `tenant_uuid`.
After that, the rest of alembic scripts is ran and everything is
migrated correctly

We need to run these scripts in pre-start because the script
`02-run-xivo-update-config` run `xivo-update-config` and it needs to
have an updated database.

We need to run wazo-update-keys before all scripts because scripts use
`wazo-upgrade` token and it must be created/updated before running
scripts
