+++ pre-start.d/000-wazo-update-keys
