+++ pre-start.d/003-load-wazo-user-from-file.py
@@ -1,18 +1,17 @@
 #!/usr/bin/env python3
-# Copyright 2018-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import argparse
 import json
 import sys
 import os
+import psycopg2
 import urllib3
 
-from contextlib import closing
-
-import psycopg2
 import requests
 
+from contextlib import closing
 from wazo_auth_client import Client as AuthClient
 from xivo.chain_map import ChainMap
 from xivo.config_helper import read_config_file_hierarchy, parse_config_file
@@ -25,7 +24,6 @@ _DEFAULT_CONFIG = {
         'key_file': '/var/lib/wazo-auth-keys/wazo-upgrade-key.yml'
     }
 }
-_DEFAULT_POLICY = 'wazo_default_admin_policy'
 
 
 def _load_config():
@@ -40,40 +38,35 @@ def _load_key_file(config):
                      'password': key_file['service_key']}}
 
 
-def _create_user(auth_client, entity_to_tenant_map, user, default_policy_uuid):
+def _create_user(auth_client, entity_to_tenant_map, user):
     tenant_uuid = entity_to_tenant_map.get(user.get('entity_id'))
+    if not tenant_uuid:
+        print('The following user has no entity, skipping... ', user)
+        return
 
     try:
-        user = auth_client.users.new(tenant_uuid=tenant_uuid, **user)
+        auth_client.users.new(tenant_uuid=tenant_uuid, **user)
     except requests.HTTPError as e:
         error = e.response.json() or {}
+        if error.get('error_id') == 'invalid-data':
+            # The email address was allowed in the php web interface, but is
+            # not allowed in wazo-auth
+            if error.get('details', {}).get('email_address', {}).get('constraint') == 'email':
+                user['email_address'] = None
+                _create_user(auth_client, entity_to_tenant_map, user)
+                return
         if error.get('error_id') == 'conflict':
             if error.get('details', {}).get('uuid', {}).get('constraint_id') == 'unique':
                 return
             elif error.get('details', {}).get('username', {}).get('constraint_id') == 'unique':
                 return
+            elif error.get('details', {}).get('email_address', {}).get('constraint_id') == 'unique':
+                return
+
         print('The user could not be migrated')
         print('The user was:', user)
         print('The error was:', error)
         raise
-
-    if not default_policy_uuid:
-        return
-
-    try:
-        auth_client.users.add_policy(user['uuid'], default_policy_uuid)
-    except requests.HTTPError as e:
-        error = e.response.json() or {}
-        print('The user could not be migrated')
-        print('The user was:', user)
-        print('The error was:', error)
-        raise
-
-
-def _find_admin_policy_uuid(auth_client):
-    policies = auth_client.policies.list(name=_DEFAULT_POLICY)['items']
-    for policy in policies:
-        return policy['uuid']
 
 
 def _build_entity_tenant_map(cursor):
@@ -85,18 +78,16 @@ def _build_entity_tenant_map(cursor):
 def _import_wazo_user(users):
     config = _load_config()
     auth_client = AuthClient(**config['auth'])
-    token = auth_client.token.new('wazo_user', expiration=36000)['token']
+    token = auth_client.token.new(expiration=36000)['token']
     auth_client.set_token(token)
 
     with closing(psycopg2.connect(config['db_uri'])) as conn:
         cursor = conn.cursor()
         entity_to_tenant_map = _build_entity_tenant_map(cursor)
 
-    default_policy_uuid = _find_admin_policy_uuid(auth_client)
-
-    print('migrating admins access to wazo-auth', end='', flush=True)
+    print('migrating users to wazo-auth', end='', flush=True)
     for user in users:
-        _create_user(auth_client, entity_to_tenant_map, user, default_policy_uuid)
+        _create_user(auth_client, entity_to_tenant_map, user)
         print('.', end='', flush=True)
     print('\ndone')
 
@@ -106,48 +97,40 @@ def _import_wazo_user(users):
 def main():
     args = parse_args()
 
-    if not args.force:
-        version_installed = os.getenv('XIVO_VERSION_INSTALLED')
-        if not version_installed:
-            print('Variable XIVO_VERSION_INSTALLED must be set')
-            sys.exit(1)
-        if version_installed > '18.14':
-            sys.exit(0)
+    if not args.force and os.getenv('XIVO_VERSION_INSTALLED') > '18.04':
+        sys.exit(0)
 
-        if version_installed <= '18.04':
-            if not os.path.exists('/var/lib/xivo-upgrade/entity_tenant_association_migration'):
-                print('22_entity_tenant_migration.py should be executed first')
-                sys.exit(1)
+    if not os.path.exists('/var/lib/xivo-upgrade/entity_tenant_migration'):
+        print('load-wazo-user-from-file: 002-create-tenants-from_entities.py should be executed first')
+        sys.exit(1)
 
-    sentinel_file = '/var/lib/xivo-upgrade/migrate_xivo_admin_to_wazo_user'
+    sentinel_file = '/var/lib/xivo-upgrade/migrate_xivo_user_to_wazo_user'
     if os.path.exists(sentinel_file):
         # migration already done
         sys.exit(1)
 
-    admin_file = '/var/lib/xivo-upgrade/xivo_admin_dump.json'
-    if not os.path.exists(admin_file):
-        print('xivo_admin_user migration failed: File {} does not exist.'.format(admin_file))
+    user_file = '/var/lib/xivo-upgrade/xivo_user_dump.json'
+    if not os.path.exists(user_file):
+        print('xivo_user to wazo_user migration failed: File {} does not exist.'.format(user_file))
         sys.exit(-1)
 
-    with open(admin_file, 'r') as f:
-        admins = json.load(f)
+    with open(user_file, 'r') as f:
+        users = json.load(f)
 
-    _import_wazo_user(admins)
+    _import_wazo_user(users)
 
     with open(sentinel_file, 'w'):
         pass
 
-    os.unlink(admin_file)
+    os.unlink(user_file)
 
 
 def parse_args():
     parser = argparse.ArgumentParser()
-    parser.add_argument(
-        '-f',
-        '--force',
-        action='store_true',
-        help="Do not check the variable XIVO_VERSION_INSTALLED. Default: %(default)s",
-    )
+    parser.add_argument('-f',
+                        '--force',
+                        action='store_true',
+                        help="Do not check the variable XIVO_VERSION_INSTALLED. Default: %(default)s")
     return parser.parse_args()
